refactor(filemanager): move pagination to its own file
