+++ lib/workload/stateless/stacks/filemanager/filemanager/src/routes/mod.rs
@@ -21,9 +21,10 @@ use crate::routes::list::*;
 use crate::routes::openapi::swagger_ui;
 
 pub mod get;
-mod ingest;
+pub mod ingest;
 pub mod list;
 pub mod openapi;
+pub mod pagination;
 
 /// App state containing database client.
 #[derive(Debug, Clone)]
